refactor module visibility
