+++ .github/workflows/workspace.yml
@@ -43,6 +43,10 @@ jobs:
         with:
           command: fmt
           args: --all -- --check
+      - uses: aig787/cargo-udeps-action@v1
+        with:
+          version: 'latest'
+          args: '--all-targets'
 
   test:
     runs-on: ubuntu-latest
