+++ ferveo/Cargo.toml
@@ -20,34 +20,24 @@ ark-poly = "0.3"
 rand = "0.8"
 rand_old = { package = "rand", version = "0.7" } # used by benchmarks/pairing.rs
 either = "1.6.1"
-hex = "0.4.2"
 miracl_core = "2.3.0"
-num = "0.4"
 ed25519-dalek = { version = "1", features = ["serde", "batch"] }
 serde = { version = "1.0", features = ["derive"] }
-blake2 = "0.10"
 blake2b_simd = "1"
 zeroize = { version = "1", default-features = false, features = ["zeroize_derive"] }
 serde_bytes = { version = "0.11" }
 bincode = "1.3"
 anyhow = "1.0"
-serde_json = "1.0"
-borsh = "0.9"
 subtle = "2.4"
 itertools = "0.10.1"
 measure_time = "0.8"
-ark-ed-on-bls12-381 = "0.3.0"
 group-threshold-cryptography = { path = "../tpke" }
 ferveo-common = { path = "../ferveo-common" }
 subproductdomain = { path = "../subproductdomain" }
 
-[dependencies.digest]
-version = "0.10.0"
-features = ["alloc"]
-
 [dev-dependencies]
 criterion = "0.3" # supports pprof, # TODO: Figure out if/how we can update to 0.4
-pprof = { version = "0.6", features = ["flamegraph", "criterion"] }
+digest = { version = "0.10.0", features = ["alloc"] }
 
 [[example]]
 name = "pvdkg"
