+++ tpke-wasm/src/lib.rs
@@ -1,12 +1,12 @@
-pub mod test;
+extern crate group_threshold_cryptography as tpke;
+
 mod utils;
 
-extern crate group_threshold_cryptography as tpke;
+use utils::*;
 
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
 use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
-use utils::set_panic_hook;
 use wasm_bindgen::prelude::*;
 
 extern crate wee_alloc;
@@ -167,4 +167,52 @@ pub fn decrypt_with_shared_secret(
 
     tpke::api::decrypt_with_shared_secret(&ciphertext.0, aad, &shared_secret.0)
         .unwrap()
+}
+
+/// Factory functions for testing
+#[cfg(any(test, feature = "test-common"))]
+pub mod test_common {
+    use super::*;
+
+    #[wasm_bindgen]
+    #[derive(Clone, Debug)]
+    pub struct Dkg {
+        pub public_key: PublicKey,
+        pub private_key: PrivateKey,
+        private_contexts: Vec<tpke::api::TpkePrivateDecryptionContext>,
+    }
+
+    #[wasm_bindgen]
+    impl Dkg {
+        #[wasm_bindgen(constructor)]
+        pub fn new(threshold: usize, shares_num: usize) -> Self {
+            set_panic_hook();
+
+            let mut rng = rand::thread_rng();
+            let (public_key, private_key, private_contexts) =
+                tpke::test_common::setup_simple::<tpke::api::E>(
+                    threshold, shares_num, &mut rng,
+                );
+            Self {
+                public_key: PublicKey(public_key),
+                private_key: PrivateKey(private_key),
+                private_contexts,
+            }
+        }
+
+        #[wasm_bindgen]
+        pub fn make_decryption_share(
+            &self,
+            ciphertext: &Ciphertext,
+            aad: &[u8],
+            validator_index: usize,
+        ) -> DecryptionShare {
+            set_panic_hook();
+            DecryptionShare(tpke::api::DecryptionShare(
+                self.private_contexts[validator_index]
+                    .create_share_precomputed(&ciphertext.0 .0, aad)
+                    .unwrap(),
+            ))
+        }
+    }
 }
