+++ tpke-wasm/tests/node.rs
@@ -3,8 +3,9 @@
 extern crate group_threshold_cryptography as tpke;
 extern crate wasm_bindgen_test;
 
-use tpke_wasm::test::*;
+use tpke_wasm::test_common::*;
 use tpke_wasm::*;
+
 use wasm_bindgen_test::*;
 
 #[test]
